+++ src/main/java/yesman/epicfight/api/ex_cap/data/modifier/WeaponModifier.java
@@ -13,15 +13,20 @@ import yesman.epicfight.api.ex_cap.managers.MovesetManager;
 import yesman.epicfight.api.ex_cap.provider.ProviderConditional;
 import yesman.epicfight.registry.deferred.holders.*;
 import yesman.epicfight.world.capabilities.item.Style;
+import yesman.epicfight.world.capabilities.item.WeaponCapability;
 import yesman.epicfight.world.capabilities.item.custom.CustomData;
 
 import java.util.Arrays;
 import java.util.List;
 import java.util.Locale;
 import java.util.Map;
+import java.util.function.Consumer;
 
 @ApiStatus.Experimental
-public record WeaponModifier(List<ResourceLocation> targets, Map<ResourceLocation, Operation> conditionalModifier, Map<Style, ResourceLocation> movesetModifier, Map<ResourceLocation, Map<DeferredCustomData<? extends CustomData<?>>, Object>> movesetCustomData, Map<DeferredCustomData<? extends CustomData<?>>, Object> weaponCustomData) {
+public record WeaponModifier(List<ResourceLocation> targets, Map<ResourceLocation, Operation> conditionalModifier, Map<Style, ResourceLocation> movesetModifier, Map<ResourceLocation, Map<DeferredCustomData<? extends CustomData<?>>, Object>> movesetCustomData, Map<DeferredCustomData<? extends CustomData<?>>, Object> weaponCustomData, Consumer<WeaponCapability.Builder> overrideModifiers) {
+    public static final Consumer<WeaponCapability.Builder> EMPTY = builder -> {};
+
+
     public enum Operation {
         APPEND,
         REMOVE
@@ -31,6 +36,7 @@ public record WeaponModifier(List<ResourceLocation> targets, Map<ResourceLocatio
         return new Builder();
     }
     public static class Builder {
+        private Consumer<WeaponCapability.Builder> builderModifiers = EMPTY;
         private final List<ResourceLocation> target;
         private final Map<ResourceLocation, Operation> conditionalModifier;
         private final Map<Style, ResourceLocation> movesetModifier;
@@ -79,6 +85,13 @@ public record WeaponModifier(List<ResourceLocation> targets, Map<ResourceLocatio
             }
             return this;
         }
+
+        public Builder modifyBuilder(Consumer<WeaponCapability.Builder> builderConsumer)
+        {
+            this.builderModifiers = builderConsumer;
+            return this;
+        }
+
 
         public void addConditionalModifier(ResourceLocation key) {
             this.conditionalModifier.put(key, Operation.APPEND);
@@ -133,6 +146,14 @@ public record WeaponModifier(List<ResourceLocation> targets, Map<ResourceLocatio
             return this.setMovesetData(moveset.getId(), data, value);
         }
 
+        /**
+         * @deprecated modifyBuilder() made this redundant
+         * @param data weapon data
+         * @param value what value
+         * @return the builder
+         * @param <T> any type
+         */
+        @Deprecated(forRemoval = true)
         public <T> Builder setWeaponData(DeferredCustomData<? extends CustomData<T>> data, T value)
         {
             this.weaponCustomData.put(data, value);
@@ -143,7 +164,7 @@ public record WeaponModifier(List<ResourceLocation> targets, Map<ResourceLocatio
             assemble(builderId);
             Map<ResourceLocation, Map<DeferredCustomData<? extends CustomData<?>>, Object>> movesetCustomData = Maps.newHashMap();
             this.movesetCustomData.forEach((style, data) -> movesetCustomData.put(style, ImmutableMap.copyOf(data)));
-            return new WeaponModifier(ImmutableList.copyOf(target), ImmutableMap.copyOf(conditionalModifier), ImmutableMap.copyOf(movesetModifier), ImmutableMap.copyOf(movesetCustomData), ImmutableMap.copyOf(weaponCustomData));
+            return new WeaponModifier(ImmutableList.copyOf(target), ImmutableMap.copyOf(conditionalModifier), ImmutableMap.copyOf(movesetModifier), ImmutableMap.copyOf(movesetCustomData), ImmutableMap.copyOf(weaponCustomData), builderModifiers);
         }
     }
 }