+++ src/main/java/yesman/epicfight/mixin/common/MixinRangedBowAttackGoal.java
@@ -1,8 +1,9 @@
 package yesman.epicfight.mixin.common;
 
+import com.llamalad7.mixinextras.injector.wrapoperation.Operation;
+import com.llamalad7.mixinextras.injector.wrapoperation.WrapOperation;
 import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.injection.At;
-import org.spongepowered.asm.mixin.injection.Redirect;
 
 import net.minecraft.world.entity.Entity;
 import net.minecraft.world.entity.LivingEntity;
@@ -13,14 +14,14 @@ import yesman.epicfight.world.capabilities.entitypatch.LivingEntityPatch;
 
 @Mixin(value = RangedBowAttackGoal.class)
 public class MixinRangedBowAttackGoal {
-	@Redirect(
+	@WrapOperation(
 		at = @At(
 			value = "INVOKE",
 			target = "Lnet/minecraft/world/entity/monster/RangedAttackMob;performRangedAttack(Lnet/minecraft/world/entity/LivingEntity;F)V"
 		),
 		method = "tick()V"
 	)
-	private void epicfight$tick(RangedAttackMob self, LivingEntity target, float velocity) {
+	private void epicfight$tick(RangedAttackMob self, LivingEntity target, float velocity, Operation<Void> operation) {
 		self.performRangedAttack(target, velocity);
 		
 		EpicFightCapabilities.getUnparameterizedEntityPatch((Entity)self, LivingEntityPatch.class).ifPresent(entitypatch -> {