+++ CHANGELOG.md
@@ -4,6 +4,27 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
+
+## [21.17.3] - 2026-5-14
+### Additions
+- Added Item Specific Overrides to Builder to allow any explicit tier configurations.
+- Added Simply Tooltips Compatibility with a plethora of Epic Fight's own features.
+### Changes
+- Used `@WrapOperation` instead of `@Redirect` to improve compatibility with other mods' mixins.
+### Deprecations
+- Deprecated `setWeaponData()` as `modifyBuilder()` made the formet method redundant.
+### Fixes
+- More inheritance issues regarding `WeaponCapability`
+- Custom Data in Item Presets not working due to logic error.
+- Skills not using `isWeaponCategory()`
+
+## [21.17.2] - 2026-05-12
+### Additions
+- Added builder modifiers to `WeaponModifier` just a small addition that gives more power to addon developers.
+### Hotfixes
+- Fixed Issue regarding dual wielding not working. Caused by a weapon category match being that just passing the entire weapon cap instead of just the category.
+- Fixed weapon category and collider inheritance by setting them to null and only adding the first categories if it stayed null from inheritance.
+
 ## [21.17.1] - 2026-04-29
 ### Changes
 **ExCap**